refactor: replace full class paths with imports for cleaner code

+++ nl2sql-core/src/main/java/com/nl2sql/core/rag/SQLFeedbackService.java
@@ -108,7 +108,7 @@ public class SQLFeedbackService {
                     try {
                         feedbackLearningService.processLowRatingFeedback(
                             feedbackId, 
-                            request.getRating(), 
+                            request.getRating().intValue(),  // Integer -> int
                             request.getQuestion(), 
                             request.getGeneratedSql(), 
                             request.getFeedbackText()
